+++ packages/geoprocessing/src/helpers/groupBy.test.ts
@@ -1,4 +1,4 @@
-import { groupBy, keyBy } from "./functions";
+import { groupBy } from "./groupBy";
 
 const list = [
   { name: "foo", id: 1 },
@@ -6,18 +6,11 @@ const list = [
   { name: "foo", id: 3 },
 ];
 
-describe("Function helper unit tests", () => {
+describe("groupBy", () => {
   test("groupBy", async () => {
     const result = groupBy(list, (item) => item.name);
     expect(Object.keys(result).length).toBe(2);
     expect(result["foo"].length).toBe(2);
     expect(result["blue"].length).toBe(1);
-  });
-
-  test("keyBy", async () => {
-    const result = keyBy(list, (item) => item.name);
-    expect(Object.keys(result).length).toBe(2);
-    expect(result.blue.id).toBe(2);
-    expect(result.foo.id).toBe(3);
   });
 });
