+++ packages/geoprocessing/src/helpers/groupBy.ts
@@ -10,13 +10,3 @@ export const groupBy = <T, K extends keyof any>(
     }
     return { ...previous, [group]: previous[group].concat(currentItem) };
   }, {} as Record<K, T[]>);
-
-/** Similar to lodash keyBy */
-export const keyBy = <T, K extends keyof any>(
-  list: T[],
-  getKey: (item: T) => K
-) =>
-  list.reduce((previous, currentItem) => {
-    const group = getKey(currentItem);
-    return { ...previous, [group]: currentItem };
-  }, {} as Record<K, T>);
